feat(ios): Support the Pimoroni Inky Impression 4 device This change also adds some drive-by fixes to ensure blank images are generated for the correct device size.
