+++ ios/StatusPanel/Extensions/UIImage.swift
@@ -55,10 +55,10 @@ extension UIImage {
     static func blankImage(size: CGSize, scale: CGFloat) -> UIImage {
         let format = UIGraphicsImageRendererFormat()
         format.scale = scale
-        let renderer = UIGraphicsImageRenderer(size: Panel.size, format: format)
+        let renderer = UIGraphicsImageRenderer(size: size, format: format)
         let image = renderer.image { context in
             UIColor.white.setFill()
-            context.fill(CGRect(origin: .zero, size: Panel.size))
+            context.fill(CGRect(origin: .zero, size: size))
         }
         return image
     }
